[AMM] Fix overflow in isAlmostEqual checks (#2062)

+++ packages/loopring_v3/contracts/amm/libamm/AmmUtil.sol
@@ -55,7 +55,7 @@ library AmmUtil
             return amount == 0;
         } else {
             // Max rounding error for a float24 is 2/100000
-            uint ratio = (amount * 100000) / targetAmount;
+            uint ratio = (uint(amount) * 100000) / uint(targetAmount);
             return (100000 - 2) <= ratio && ratio <= (100000 + 2);
         }
     }
@@ -72,7 +72,7 @@ library AmmUtil
             return amount == 0;
         } else {
             // Max rounding error for a float16 is 5/1000
-            uint ratio = (amount * 1000) / targetAmount;
+            uint ratio = (uint(amount) * 1000) / uint(targetAmount);
             return (1000 - 5) <= ratio && ratio <= (1000 + 5);
         }
     }
